+++ R/lsos.R
@@ -13,8 +13,9 @@
 #' details.
 
 #' @author 
-#' Both functions were created by Dirk Edelbuettel and modified by JD Long. 
-#' See \url{http://stackoverflow.com/questions/1358003/} for details.
+#' Both functions were created by Dirk Edelbuettel and modified by JD Long 
+#' and John Bullock. See \url{http://stackoverflow.com/questions/1358003/} 
+#' for details.
 #' 
 #' @return 
 #' Data frame with columns "Class," "Size," "Row," and "Columns." The row 
@@ -22,6 +23,8 @@
 #' 
 #' 
 #' @param ... Additional arguments to \code{.ls.objects()}.
+#' @param MB Logical. If \code{TRUE} (the default), object size is reported 
+#' in megabytes. If \code{FALSE}, it is reported in kilobytes.
 #' @param pos Numeric. Specifies the position, in the search list, of the
 #' environment to search. Can be specified instead of \code{envir}.
 #' @param envir Environment to search. Can be specified instead of \code{pos}.
@@ -44,7 +47,7 @@
 
 #' @rdname lsos
 .ls.objects <- function (pos = 1, envir = as.environment(pos), pattern, order.by,
-                        decreasing=FALSE, n=NULL) {
+                        decreasing=FALSE, MB = TRUE, n=NULL) {
     napply <- function(names, fn) sapply(names, function(x)
                                          fn(get(x, pos = pos)))
     names <- ls(envir = envir, pattern = pattern)
@@ -55,18 +58,24 @@
     obj.class <- napply(names, function(x) as.character(class(x))[1])
     obj.mode <- napply(names, mode)
     obj.type <- ifelse(is.na(obj.class), obj.mode, obj.class)
-    obj.size <- napply(names, utils::object.size)
+    obj.size <- napply(
+        names, 
+        function (x) format(utils::object.size(x), units = if (MB) "MB" else "Kb", digits = 3)) %>%
+      { gsub("^([\\d\\.]*).*", "\\1", .data, perl = TRUE) } %>%  # extract numbers from "0.005 Mb", etc.
+      as.numeric()
+    
     obj.prettysize <- sapply(obj.size, function(r) prettyNum(r, big.mark = ",") )
-    obj.dim <- t(napply(names, function(x)
-                        as.numeric(dim(x))[1:2]))
+    obj.dim <- t(napply(names, function(x) as.numeric(dim(x))[1:2]))
     vec <- is.na(obj.dim)[, 1] & (obj.type != "function")
     obj.dim[vec, 1] <- napply(names, length)[vec]
+    
     out <- data.frame(obj.type, obj.size, obj.prettysize, obj.dim)
     names(out) <- c("Class", "Size", "PrettySize", "Rows", "Columns")
     if (!missing(order.by))
         out <- out[order(out[[order.by]], decreasing=decreasing), ]
         out <- out[c("Class", "PrettySize", "Rows", "Columns")]
         names(out) <- c("Class", "Size", "Rows", "Columns")
+    if (MB) names(out)[2] <- "Size (MB)" else "Size (Kb)"
     if (!is.null(n))
         out <- utils::head(out, n)
     out
@@ -75,7 +84,7 @@
 
 #' @rdname lsos
 #' @export
-lsos <- function(..., n = 8) {
-    .ls.objects(..., order.by="Size", decreasing=TRUE, n=n)
+lsos <- function(..., MB = TRUE, n = 8) {
+    .ls.objects(..., MB = TRUE, order.by="Size", decreasing=TRUE, n=n)
 }